send to git

+++ handle_pointer.c
@@ -1,3 +1,14 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   handle_pointer.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: fivieira <fivieira@student.42porto.com>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/05/18 09:45:30 by fivieira          #+#    #+#             */
+/*   Updated: 2023/05/18 10:20:12 by fivieira         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "ft_printf.h"
 
@@ -11,11 +22,11 @@ static int	print_string(char *str)
 
 int	handle_pointer(void *ptr_addr)
 {
-	int		counter;
-	char	*str;
-	unsigned int addr;
+	int				counter;
+	char			*str;
+	unsigned long	addr;
 
-	addr = (unsigned int)ptr_addr;
+	addr = (unsigned long)ptr_addr;
 	counter = print_string(PRECEDING_STR);
 	str = ft_itoa(addr);
 	counter = counter + handle_string(str);
